- Sudden Death Bug Fixed - You can now kill Aliens when stepping on his head - Multiplier coming Soon - Platforms now grant you points - Huehuheue

+++ public_html/js/play.js
@@ -2,7 +2,7 @@ Game.Play = function() {
     this.currentState;
     this.moving;
     this.createStage = function(width, height, background, platform) {
-
+        
         game.world.setBounds(0, 0, width, height);
 
         for (var i = 0; i < width; i = i + 800) {
@@ -40,18 +40,18 @@ Game.Play = function() {
     this.createEnemies = function(number, width, sprite) {
         this.enemies = game.add.group();
         for (var i = 0; i < number; i++) {
-            var enemy = this.enemies.create((Math.random() * width) + 100, Math.random() * 300, sprite);
-            enemy.body.bounce.y =31.0;
-          //  enemy.body.collideWorldBounds = true;
-           // enemy.body.gravity.y = -10;
-          //  enemy.body.acceleration.y = 400;
-         /*               game.add.tween(enemy)
-                    .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
-                    .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
-                    .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
-                    .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
-                    .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
-                    .start();*/
+            var enemy = this.enemies.create((Math.random() * width) + 300, Math.random() * 300, sprite);
+            enemy.body.bounce.y = 31.0;
+            //  enemy.body.collideWorldBounds = true;
+            // enemy.body.gravity.y = -10;
+            //  enemy.body.acceleration.y = 400;
+            /*               game.add.tween(enemy)
+             .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
+             .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
+             .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
+             .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
+             .to({x: Math.random() * 3100}, 1000, Phaser.Easing.Linear.None)
+             .start();*/
         }
     };
 
@@ -133,6 +133,7 @@ Game.Play = function() {
 
         this.ledges.setAll('body.immovable', true);
         this.ledges.setAll('body.collideWorldBounds', true);
+        this.ledges.setAll('health', 2);
     };
 
 
@@ -162,7 +163,17 @@ Game.Play = function() {
         }
     };
 
-    this.killPlayer = function() {
+    this.killPlayer = function(player,enemy) {
+       if (this.player.body.touching.down && enemy.body.touching.up)
+            {
+                enemy.kill();
+                this.enemies.remove(enemy);
+                this.enemies.create((Math.random() * 2990) + 100, Math.random() * 300, 'enemy');
+                this.collectText("+10");
+                time += 10 * mult;
+                return;
+            }
+       
         this.player.alive = false;
         this.player.kill();
     };
@@ -172,14 +183,23 @@ Game.Play = function() {
         this.enemies.getRandom().kill();
     };
 
-
+    this.platformPoints = function(player, ledger){
+      if((this.player.body.touching.down) && (ledger.body.touching.up) && (ledger.health === 2)){
+            ledger.health = 1;
+            var temp = ((this.roundTo10(game.world.height - this.player.y) / 10) / 2) * mult;
+            var temp = parseInt(temp);
+            this.collectText(temp);
+            time += temp;
+      };  
+    };
+    
     this.killAllEnemies = function() {
         this.enemies.callAll('kill');
     };
 
     this.createScore = function(text, x, y) {
-        game.context.font = "40px Arial";
-        game.context.fillStyle = "Black";
+        game.context.font = "25px Arial";
+        game.context.fillStyle = "White";
         game.context.fillText(text, x, y);
     };
 
@@ -196,6 +216,27 @@ Game.Play = function() {
     };
 
 
+    this.collectText = function(text) {
+        var style = {font: "10pt Arial", fill: "#ffffff", align: "center"};
+        var textCollect = game.add.text(this.player.x, this.player.y - 10, text, style);
+        this.texts.add(textCollect);
+        textCollect.alpha = 0;
+        game.add.tween(textCollect)
+                .to({alpha: 1, y: this.player.y - 40}, 400, Phaser.Easing.Linear.None, true)
+                .onComplete.add(this.clearCollect, this);
+        
+    };
+    
+    this.roundTo10 = function(number){
+        var temp = number / 10;
+        Math.round(temp);
+        temp *=temp * 10;
+        return temp;
+    };
+    
+    this.clearCollect = function() {
+        this.texts.removeAll();
+    };
 
     this.playerMovement = function() {
         this.player.body.velocity.x = 0;
@@ -229,7 +270,7 @@ Game.Play = function() {
         if (cursors.down.isDown)
         {
             this.player.body.velocity.y = 700;
-            this.moving = true;
+            this.moving = false;
         }
     };
